zetaclient implementation
